TextContentTagElement replaces TextTagElement
(Heading|Paragraph|Component)TagElement classes (still) extend TextTagElement, wich is now what was TextContentTagElement.
All three also re-parse the "shadow" parameter, if passed, defaulting to "false".

- HeadingTagElement keeps it's overrides, as they account for the scaling used to render H1 and H2.
- ComponentTagElement reduces to a single override of `addText()` to parse the json content.
- ParagraphTagElement class reduces to just it's initializer.

- DefaultTagProvider uses the new name, "TextTagElement", to for the serializel for `<text>` elements.
