+++ common/src/main/java/earth/terrarium/hermes/api/defaults/TextTagElement.java
@@ -3,69 +3,124 @@ package earth.terrarium.hermes.api.defaults;
 import com.teamresourceful.resourcefullib.common.color.Color;
 import earth.terrarium.hermes.api.Alignment;
 import earth.terrarium.hermes.api.TagElement;
+import earth.terrarium.hermes.api.TagProvider;
+import earth.terrarium.hermes.api.text.ChildTextTagElement;
+import earth.terrarium.hermes.api.text.TextTagProvider;
+import earth.terrarium.hermes.api.themes.Theme;
 import earth.terrarium.hermes.utils.ElementParsingUtils;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.Style;
 import net.minecraft.util.FormattedCharSequence;
-import org.jetbrains.annotations.Nullable;
+import net.minecraft.util.Mth;
 
+import java.util.List;
 import java.util.Map;
 
-public abstract class TextTagElement extends FillAndBorderElement implements TagElement {
+public class TextTagElement extends FillAndBorderElement implements TagElement {
 
-    protected String content = "";
-    protected @Nullable Boolean bold;
-    protected @Nullable Boolean italic;
-    protected @Nullable Boolean underline;
-    protected @Nullable Boolean strikethrough;
-    protected @Nullable Boolean obfuscated;
-    @Deprecated protected boolean centered;
+    protected MutableComponent component = Component.empty();
     protected Alignment align;
     protected boolean shadowed;
-    protected Color color;
 
-    protected TextTagElement(Map<String, String> parameters) {
+    public TextTagElement(Map<String, String> parameters) {
         super(parameters);
-        this.bold = parameters.containsKey("bold") ? Boolean.parseBoolean(parameters.get("bold")) : null;
-        this.italic = parameters.containsKey("italic") ? Boolean.parseBoolean(parameters.get("italic")) : null;
-        this.underline = parameters.containsKey("underline") ? Boolean.parseBoolean(parameters.get("underline")) : null;
-        this.strikethrough = parameters.containsKey("strikethrough") ? Boolean.parseBoolean(parameters.get("strikethrough")) : null;
-        this.obfuscated = parameters.containsKey("obfuscated") ? Boolean.parseBoolean(parameters.get("obfuscated")) : null;
-        // 'centered' is DEPRECATED
-        this.centered = parameters.containsKey("centered") && Boolean.parseBoolean(parameters.get("centered"));
-        this.shadowed = parameters.containsKey("shadowed") && Boolean.parseBoolean(parameters.get("shadowed"));
+        this.component.setStyle(Style.EMPTY
+                .withBold(ElementParsingUtils.parseBoolean(parameters, "bold", false))
+                .withItalic(ElementParsingUtils.parseBoolean(parameters, "italic", false))
+                .withUnderlined(ElementParsingUtils.parseBoolean(parameters, "underline", false))
+                .withObfuscated(ElementParsingUtils.parseBoolean(parameters, "obfuscated", false))
+                .withStrikethrough(ElementParsingUtils.parseBoolean(parameters, "strikethrough", false))
+                .withColor(ElementParsingUtils.parseColor(parameters, "color", Color.DEFAULT).getValue())
+        );
         this.align = ElementParsingUtils.parseAlignment(parameters, "align", Alignment.MIN);
-        if (parameters.containsKey("color")) {
-            try {
-                this.color = Color.parse(parameters.get("color"));
-            } catch (Exception ignored) {
-                this.color = Color.DEFAULT;
+        this.shadowed = ElementParsingUtils.parseBoolean(parameters, "shadowed", true);
+    }
+
+    @Override
+    public void render(Theme theme, GuiGraphics graphics, int x, int y, int width, int mouseX, int mouseY, boolean hovered, float partialTicks) {
+        x = x + xSurround;
+        y = y + ySurround;
+
+        Font font = Minecraft.getInstance().font;
+        List<FormattedCharSequence> lines = font.split(component, width - (5 + (2 * xSurround)));
+        int maxWidth = lines.stream().mapToInt(font::width).max().orElse(0);
+        int maxHeight = (lines.size() * font.lineHeight) + (lines.size() - 2);
+        int offsetX = Alignment.getOffset(width, maxWidth + (2 * xSurround), align);
+
+        drawFillAndBorder(graphics, x + offsetX, y, maxWidth, maxHeight);
+
+        int actMouseX = mouseX - x;
+        int actMouseY = mouseY - y;
+        int height = 0;
+        for (FormattedCharSequence sequence : font.split(component, width - (5 + 2 * xSurround))) {
+            int textOffset = getOffsetForTextTag(width, sequence);
+            theme.drawText(graphics, sequence, x + textOffset, y + height, Color.DEFAULT, this.shadowed);
+
+            if (actMouseX >= textOffset && actMouseX <= width && actMouseY >= height && actMouseY <= height + font.lineHeight) {
+                graphics.renderComponentHoverEffect(
+                    font,
+                    font.getSplitter().componentStyleAtWidth(sequence, Mth.floor(actMouseX - textOffset)),
+                    mouseX, mouseY
+                );
             }
-        } else {
-            this.color = Color.DEFAULT;
-        }
-        // 'centered' is now DEPRECATED; for now, 'centered="true"' parameter over-rides the align parameter
-        if (this.centered) {
-            align = Alignment.MIDDLE;
+
+            height += Minecraft.getInstance().font.lineHeight + 1;
         }
     }
 
     @Override
-    public void setContent(String content) {
-        this.content = content;
+    public boolean mouseClicked(double mouseX, double mouseY, int button, int width) {
+        Font font = Minecraft.getInstance().font;
+        int height = 0;
+        for (FormattedCharSequence sequence : font.split(component, width - 5)) {
+            int textOffset = getOffsetForTextTag(width, sequence);
+            if (mouseX >= textOffset && mouseX <= width && mouseY >= height && mouseY <= height + font.lineHeight) {
+                Style style = font.getSplitter().componentStyleAtWidth(sequence, Mth.floor(mouseX - textOffset));
+                if (Minecraft.getInstance().screen != null) {
+                    Minecraft.getInstance().screen.handleComponentClicked(style);
+                }
+                return true;
+            }
+            height += Minecraft.getInstance().font.lineHeight + 1;
+        }
+        return false;
+    }
+
+    @Override
+    public int getHeight(int width) {
+        Font font = Minecraft.getInstance().font;
+        int lines = font.split(component, width - (5 + (2 * xSurround))).size();
+        int lineHeight = font.lineHeight;
+        return ((lines * lineHeight) + (lines - 2)) + (2 * ySurround);
+    }
+
+    @Override
+    public void addText(String content) {
+        this.component.append(content);
+    }
+
+    @Override
+    public void addChild(TagElement element) {
+        if (element instanceof TextTagElement textTag) {
+            this.component.append(textTag.component);
+        } else if (element instanceof ChildTextTagElement textTag) {
+            this.component.append(textTag.component());
+        } else {
+            throw new UnsupportedOperationException();
+        }
+    }
+
+    @Override
+    public TagProvider getChildTagProvider(TagProvider parent) {
+        return TextTagProvider.INSTANCE;
     }
 
     public int getOffsetForTextTag(int width, FormattedCharSequence text) {
         int textWidth = Minecraft.getInstance().font.width(text) - 1; // -1 to trim trailing empty space
         return Alignment.getOffset(width, textWidth + (2 * xSurround), align);
-    }
-
-    public Style getStyle() {
-        return Style.EMPTY
-                .withBold(bold)
-                .withItalic(italic)
-                .withUnderlined(underline)
-                .withStrikethrough(strikethrough)
-                .withObfuscated(obfuscated);
     }
 }
